requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sentence-transformers
faiss-cpu
numpy
pandas
requests
